prepare data at PAGE
